Removed the yield from Channel#exchange, and removed bad tests
The tests were bad because 1) they didn't fail when the yield was removed and 2) they checked the concrete type of the yielded objects. 

Also, the DeclareOk and DeleteOk should not be allowed to escape to the client, they are part of the low level Java API and don't belong in the Hot Bunnies high level world.

DeleteOk is returned from #delete, but that is just an artifact of Ruby, I would prefer not to make it a part of the API, and having tests that check the type of the returned object would require us to support that in future versions.