Rewrote the internals of Queue#subscribe
The old subscription code used com.rabbitmq.QueueingConsumer, which has been deprecated. It would also block indefinitely when there were no more messages in the queue, and the executor that ran it had to be force shut down.

The new code also runs each delivery in a separate task, which means that a single delivery raising an error no longer kills the whole delivery pipe. A test was changed to reflect this.