Introduce Channel#ack, #reject and #nack
The idea is to move the API as close to Bunny 0.9 as we can,
while still being backwards compatible as much as we can.

Failure scenarios require Java exceptions to be caught,
converted to Ruby exceptions and re-thrown, otherwise
it's not possible for Ruby developers to make sense of
how to catch them and what information they can obtain
from them.