Selecting the right overloaded Executor#submit
JRuby 1.7.x is more aggressive in warning when there is not enough type information and it has to pick an overloaded implementation. This avoids the warning by selecting the method explicitly. Could have used #java_send too, but that would be noisier. Executor#execute works too, but it prints warnings in the tests that I don't want to have to understand why right now.

This fixes #14