Make BlockingConsumer#cancel idempotent
Make sure that Channel#basic_cancel is only called once for the consumer tag. As it is a raw RMQ exception is raised if you accidentally cancel twice (that's also the reason for the `rescue NativeException` in the specs, it makes the stack trace come out correctly when the test fails).