Move #cancel implementation from BlockingConsumer to CallbackConsumer
Also change its semantics slightly: it now returns true if it did anything, and false if it did nothing (i.e. the consumer had already been cancelled).

The implementation in CallbackConsumer is wrong, it does not flip the @cancelling, @cancelled and @terminated flags (which are defined by BaseConsumer, so maybe this should be moved up one more notch), and it does not guard against double cancel.

The three consumer classes should probably be overhauled. Maybe there should be only one now that we don't have any difference between blocking and sync.

Also, the enqueueing of the poison pill in BlockingConsumer should maybe be done in #handleCancel instead of #cancel, because messages can still be delivered after a cancel, can't they?