Introduce :thread_pool_size instead of an executor factory
We cannot use executors because they cannot be restarted but
judging from the feedback for 2.0, people don't get why
they need to pass in a factory, and also want a fixed
size executor most of the time.

This option constructs a callable that always returns
an executor of the provided size.