Add modern rspec configuration generated with rspec --init
Modern rspec versions include a lot of options that are recommended but
not on by default so as to not break backwards-compatibility with users
upgrading from earlier 3.x versions.

When generating a new spec_helper with rspec --init, rspec sets a number
of these recommended defaults, of which I enabled all but the warnings
and the disable_monkey_patching!