Support creating thread factories & pools that use daemon threads
Java/JVM threads introduce the concept of daemon threads: any thread can
either be non-daemon or daemon.

As discussed in
https://stackoverflow.com/questions/2213340/what-is-daemon-thread-in-java
the difference between both is only at JVM exit: the JVM exits when all
non-daemon threads terminate. When the last non-daemon thread finishes
its work and terminates, the JVM will shut down, and any remaining
daemon threads are just ignored.

By default, march_hare either relies on the ampq-client to create its
own threads and pools or it supplies its own pools. In both of these
cases, all the created threads are non-daemon, that is, they block the
JVM exit until they finish. This means that if you start a march_hare
connection or consumer, and don't call #close on it, the JVM won't exit,
even if there's nothing else to do.

On the Ruby side, the concept of daemon/non-daemon threads does not
exist. On MRI, when the first thread created on the system dies, the
system exits.

This creates an expectation mismatch when rubyists start using
march_hare: suddenly their applications just "hang" when exiting, without
any clear reason why.

This happens a lot on @Talkdesk as we move more and more of our systems
from bunny to march_hare: in many cases, our connections are long-lived
singletons and thus we never closed them, but suddenly things like our
test suites or our rake tasks just "hang" if for some reason they cause
a rabbimq connection to be started but just throw away the reference
without closing the connection when they don't need it anymore.

This issue is so common that it made it to our internal JRuby FAQ. I've
also seen it in questions such as
https://github.com/ruby-amqp/march_hare/issues/97#issuecomment-302388621

To facilitate the use of march_hare using daemon threads instead of
the default non-daemon threads, this commit adds the option of creating
thread factories and pools that make use of daemon threads to the
MarchHare::ThreadPools class.

To configure MarchHare to use daemon threads, you can then do:

```ruby
require 'march_hare'

connection = MarchHare.connect(
  uri: 'amqp://guest:guest@localhost',
  thread_factory: MarchHare::ThreadPools.daemon_thread_factory,
  executor_factory: proc {
    MarchHare::ThreadPools.fixed_of_size(10, use_daemon_threads: true)
  },
)
```

The `thread_factory` is used by the underlying Java client to create
new threads to handle connections, and the optional `executor_factory`
can be used to get even more control on the number and allocation rules
used by the thread pool.