RFC: Configure all background threads as daemon threads by default
Java/JVM threads introduce the concept of daemon threads: any thread can
either be non-daemon or daemon.

As discussed in
https://stackoverflow.com/questions/2213340/what-is-daemon-thread-in-java
the difference between both is only at JVM exit: the JVM exits when all
non-daemon threads terminate. When the last non-daemon thread finishes
its work and terminates, the JVM will shut down, and any remaining
daemon threads are just ignored.

By default, march_hare either relies on the ampq-client to create its
own threads and pools or it supplies its own pools. In both of these
cases, all the created threads are non-daemon, that is, they block the
JVM exit until they finish. This means that if you start a march_hare
connection or consumer, and don't call #close on it, the JVM won't exit,
even if there's nothing else to do.

On the Ruby side, the concept of daemon/non-daemon threads does not
exist. On MRI, when the first thread created on the system dies, the
system exits.

This creates an expectation mismatch when rubyists start using
march_hare: suddenly their applications just "hang" when exiting, without
any clear reason why.

This happens a lot on @Talkdesk as we move more and more of our systems
from bunny to march_hare: in many cases, our connections are long-lived
singletons and thus we never closed them, but suddenly things like our
test suites or our rake tasks just "hang" if for some reason they cause
a rabbimq connection to be started but just throw away the reference
without closing the connection when they don't need it anymore.

This issue is so common that it made it to our internal JRuby FAQ. I've
also seen it in questions such as
https://github.com/ruby-amqp/march_hare/issues/97#issuecomment-302388621

To solve this challenge, this PR changes march_hare to behave like
MRI/bunny, and use daemon threads by default for all background work.
This way examples such as:

```ruby
require 'march_hare'

con = MarchHare.connect(
    automatic_recovery: false,
    uri: 'amqp://guest:guest@localhost',
)
```

no longer hang the JVM, and instead exit cleanly.

Users that want the existing behavior can still supply their own thread
pools as they've always could, and the defaults for manually-created
MarchHare#ThreadPools are also unchanged.

I'm opening this as a "Request for Comments" as this still changes the
default behavior, but I really believe this will save a lot of
head-scratching for a lot of library users.