Default to use TLS verification when certificate path and pass is set.
When a TLS/SSL certificate path and password is provided for a PKCS12 keystore, those certificates from that keystore will now also be used (by default) for the TLS/SSL trust instead of the `NullTrustManager`.
`options[:trust_manager]` is now available to be set when certificate path and password are provided for a PKCS12 keystore.

Fixes: #127